replace testCellToChildren with testCellToChildren_new
